Fix code.asm parsing on Windows
- support parsing of addresses printed using %p format: on Windows its
[A-F0-9]{8} and [A-F0-9]{16} rather then 0x[a-f0-9]{8} and 0x[a-f0-9]{16}
printed on Linux and Mac;
- remove line ending normalization, V8 has been correctly opening code trace
files in binary mode since February 2015 - so original line endings should be
correctly retained in the code.asm output. This effectively means that we are
dropping support for compilation artifacts produced by old versions of node on
Windows (which is arguably not a big loss).